Relaxing fsm-state-activity-do argument constraint
The arg0 argument was enforced in the XML schema to be equal with the
name of a defined link. However, the arg0 can indicate anything
logically. So this constraint is removed from the XML schema.
Anyway, the code (during instantiating the map) checks this constraint
for link-down and link-up commands.