Making Simulator thread's methods non-blocking
Methods who cannot acquire locks, either allow other slots to be called
via manual event processing or will return to be invoked later instead
of blocking for the lock.

This should work efficiently because:
  1. Only two threads (simulator and map) compete for locks,
  2. Situations that both threads seek locks are rare; e.g. when
  Simulator wants to send a new frame while map's thread sends a new
  frame for it.

Also some consistency checks are added. This is essential to prevent
Simulator malfunction when Simulator threads get merged into the main
thread as malformed packets must not be able to make any slot to block
for data.